requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
streamlit>=1.32
chromadb>=0.5
python-dotenv>=1.0

# Optional (choose one embedding provider)
sentence-transformers>=2.6
torch>=2.1

# Optional (LLM)
openai>=1.30
